requirements.txt
python-telegram-bot[job-queue]==20.7
pytz==2024.1
APScheduler==3.10.4
setuptools==80.0.0